mini_sendmail.c: Make -p and -s flags work again.
We have broken the arguments parser somewhere in the past and it was not
possible to specify either of these two options despite compiling with
DO_MINUS_SP defined.

Also added these options to the message printed when --help was issued.

version.h: Changed to msmtp/0.4.10 22.Mar.2012 due to the bugfix.